requirements.txt
Flask==2.3.3
python-dotenv==1.0.1
youtube-transcript-api==0.6.1
requests==2.31.0
